# Bump merkle-tree - v0.2.0