--- a/go.mod
+++ b/go.mod
@@ -15,7 +15,7 @@ require (
 	github.com/segmentio/golines v0.11.0
 	github.com/spacemeshos/api/release/go v1.5.6
 	github.com/spacemeshos/go-scale v1.1.3
-	github.com/spacemeshos/merkle-tree v0.1.0
+	github.com/spacemeshos/merkle-tree v0.2.0
 	github.com/stretchr/testify v1.8.1
 	github.com/syndtr/goleveldb v1.0.1-0.20210819022825-2ae1ddf74ef7
 	go.uber.org/zap v1.24.0