--- a/hash/hash.go
+++ b/hash/hash.go
@@ -2,6 +2,7 @@ package hash
 
 import (
 	"github.com/minio/sha256-simd"
+	"github.com/spacemeshos/merkle-tree"
 )
 
 // LabelHashNestingDepth is the number of recursive hashes per label.
@@ -12,32 +13,33 @@ const LabelHashNestingDepth = 100
 //
 // ⚠️ The resulting function is NOT thread-safe, however different generated instances are independent.
 // The code is optimized for performance and memory allocations.
-func GenMerkleHashFunc(challenge []byte) func(lChild, rChild []byte) []byte {
-	var buffer []byte
-	return func(lChild, rChild []byte) []byte {
-		size := len(challenge) + len(lChild) + len(rChild)
-		if len(buffer) < size {
-			buffer = make([]byte, size)
-		}
-		copy(buffer, challenge)
-		copy(buffer[len(challenge):], lChild)
-		copy(buffer[len(challenge)+len(lChild):], rChild)
-
-		result := sha256.Sum256(buffer[:size])
-		return result[:]
+func GenMerkleHashFunc(challenge []byte) merkle.HashFunc {
+	h := sha256.New()
+	return func(buf, lChild, rChild []byte) []byte {
+		h.Reset()
+		_, _ = h.Write(challenge)
+		_, _ = h.Write(lChild)
+		_, _ = h.Write(rChild)
+		return h.Sum(buf)
 	}
 }
 
 // GenLabelHashFunc generates hash functions for computing labels. The challenge is prepended to the data and the result
 // is hashed using Sha256. TODO: use nested hashes based on a difficulty param.
 func GenLabelHashFunc(challenge []byte) func(data []byte) []byte {
+	var hashBuf [sha256.Size]byte
+	h := sha256.New()
 	return func(data []byte) []byte {
-		message := append(challenge, data...)
-		var res [32]byte
-		for i := 0; i < LabelHashNestingDepth; i++ {
-			res = sha256.Sum256(message)
-			message = res[:]
+		h.Reset()
+		_, _ = h.Write(challenge)
+		_, _ = h.Write(data)
+		hash := h.Sum(hashBuf[:0])
+
+		for i := 1; i < LabelHashNestingDepth; i++ {
+			h.Reset()
+			_, _ = h.Write(hash)
+			hash = h.Sum(hash[:0])
 		}
-		return message
+		return hash[:]
 	}
 }