--- a/verifier/verifier.go
+++ b/verifier/verifier.go
@@ -14,7 +14,7 @@ import (
 // leaves matches the security param, validates the Merkle proof itself and verifies the labels are derived from the
 // left cousins in the Merkle tree.
 func Validate(proof shared.MerkleProof, labelHashFunc func(data []byte) []byte,
-	merkleHashFunc func(lChild, rChild []byte) []byte, numLeaves uint64, securityParam uint8,
+	merkleHashFunc merkle.HashFunc, numLeaves uint64, securityParam uint8,
 ) error {
 	if int(securityParam) != len(proof.ProvenLeaves) {
 		return fmt.Errorf("number of proven leaves (%d) must be equal to security param (%d)",