Before:
```
BenchmarkGetProof-20    	       1	30006684789 ns/op	   5.30 MB/s	   4965833 leafs/op	    165528 leafs/sec	10519326320 B/op	44632446 allocs/op
```

After:
```
BenchmarkGetProof-20    	       1	30029381371 ns/op	   5.54 MB/s	   5195681 leafs/op	    173189 leafs/sec	977521384 B/op	    6326 allocs/op
```